[fix]: casting and formater

+++ 0x1E-search_algorithms/0-linear.c
@@ -17,9 +17,9 @@ int linear_search(int *array, size_t size, int value)
 	for (i = 0; i < size; i++)
 	{
 		if (array[i] != value)
-			printf("Value checked array[%d] = %d\n", (int)i, array[i]);
+			printf("Value checked array[%lu] = %d\n", i, array[i]);
 		else
-			return ((int)i);
+			return (i);
 	}
 	return (-1);
 }
